Use exponential fade-in/out to avoid nasty clicks
Step changes in waveforms (e.g. if you suddenly start or stop one
mid-phase) create high-frequency artifacts and are audible as clicks. By
ramping to the desired value we can eliminate these.

I'm not sure this is really needed for the start of a note since if it
always starts at the zero-point, there'd be no discontinuity. But we'll
need to look at attack later so why not.