Use Hammond to generate manual keys
This makes the Hammond class responsible for knowing what keys it
supports on the manuals, and to do so it can just index into its
tonewheels instead of doing funky octave looping stuff.

I seem to be going back and forth on how to refer to the identifiers "C"
and "A#" etc. and so forth.

I'm led to believe that e.g. "C2" is a pitch, and "C" is a "pitch
class", but "pitch class" feels unwieldy to me even if technically
correct. At the moment I seem to be settling on "key" to refer to it
when it maps to a physical key, and "tone" when it refers to a
tonewheel. But there's not much rationale there. Maybe I'll give in and
go with "pitch class" at a later date.