Enable multiple simultaneous oscillators
Now we've got a single key playing a single tone, we need to move on to
supporting polyphony and drawbars, both of which involve playing
multiple tones at once.

This rejigs the synth so that all oscillators are started on setup, and
have their own gain connected to the output.

We maintain a set of currently-pressed keys, and on keychange first
modify this set, and then update the gains for all tonewheels. Currently
we just set the gain for the fundamental tone for each key to exactly
1.0.

This automatically supports polyphony (since our currently-pressed set
of keys can contain multiple entries, even if we don't yet have the
UI to actually press multiple keys), and provides an easy entrypoint for
implementing drawbars, since we can permit each pressed key to modify
the gains of multiple tonewheels.

             +----------+ +---------+
             | Drawbars | |  Keys   |
             +-----+----+ +----+----+
                   |           |
                   +-----+-----+
                         |
                  +------v------+
    +---------+   | +---------+ |
    |  Wheel  +----->  Gain   +-----+
    +---------+   | +---------+ |   |
                  |             |   |
    +---------+   | +---------+ |   |      +---------+
    |  Wheel  +----->  Gain   +-----+------>   Out   |
    +---------+   | +---------+ |   |      +---------+
                  |             |   |
    +---------+   | +---------+ |   |
    |  Wheel  +----->  Gain   +-----+
    +---------+   | +---------+ |
                  +-------------+