Pull out all the stops!
This implements the full voicing of a pressed key, with all 9 harmonics
hard-coded to the max value for now.

I started getting nasty distortion so I've turned the main gain node
down which seemed to get rid of it; I'm still not really sure what I'm
doing with the overall gain of the system, or how it'll work when I need
to combine multiple notes.

Harmonics sourced from [1], and note that as with the real Hammond,
the second "drawbar" is located below the fundamental, even though it
plays a higher tone.