Switch to midi notes for synth interface
As noted previously I was getting in a naming mess with keys and notes
and tones and wotnot. This was also tangled up in calculating weird
offsets into the tonewheel index.

I'm therefore switching to using midi note numbers[1,2] as the standard
interface between controls (whether mouse or keyboard) and the synth
itself.

[1] https://en.wikipedia.org/wiki/MIDI#:~:text=MIDI%20notes%20are%20numbered%20from,A0%20to%20C8.
[2] https://www.inspiredacoustics.com/en/MIDI_note_numbers_and_center_frequencies