Static, non-functional drawbars
This is a first pass at implementing a slider control that looks
passably like a Hammond drawbar, using only HTML and CSS. They consist
of 8 value segments and a thumb, pulling towards the viewer (down on the
page) to increase the value.

I'm quite pleased that the Cabin webfont appears to be a decent match
for the typeface on the Hammond itself.