Scrollable drawbar element
To make the drawbar into a usable slider control I want to use HTML
scrolling, controlled by JS to manage it in discrete steps corresponding
to each value notch.

This demonstrates using scroll: hidden and a bit of bounding box
calculation to manage the scroll increments. We need to add the right
amount of overflowing content below the thumb to give the whole drawbar
the right range of motion; I've bodged it in here with an invisible set
of extra track segments, but I can probably just do it with a bottom
margin on the thumb.