Allow stepped dragging of the drawbar
The main functional change here is to implement stepped dragging of the
drawbar. I've introduced a Drawbar class that's now responsible for
building the actual drawbar template when you give it a container
element.

I've also dispensed with the invisible track segments from the previous
commit and used a calculated bottom margin on the thumb to implement the
drawbar overflow.