Extract top-level playMidiNote and stopMidiNote
We've got a bit of duplicated handling of the "pressed" state for the
visual key representations, because we've got handlers for both clicks
in the UI and keypresses on the computer keyboard.

I'm about to add a third input source (MIDI inputs!) so I want a
generalised way to start and stop playing a note.

I think this points to a wider refactor I need to do, something along
the lines of introducing a top-level Hammond object whose job is to
coordinate the underlying synth, the visualisation and the various
inputs.