Bodge in MIDI control
It works! I can play all the notes at once!

It's a bit wonky because currently if a MIDI key is the first thing I
press, it fails to initialise webaudio (because that's not considered a
user input for the purposes of enabling sound), but my initialisation
code gets stuck so I need to reload the page and make sure I click or
press a (computer) keyboard key first.

Code cribbed from the MDN examples.

https://developer.mozilla.org/en-US/docs/Web/API/Web_MIDI_API