Extract MidiHammondInput object
This un-bodges the MIDI input introduced previously, and starts a
pattern for Hammond I/O. We've got a number of different components now:

* Synth (currently the Hammond object): output only
* HTML UI: input and (visual) output
* Computer keyboard: input only
* MIDI: input only

I'd like to establish a top-level controller object that can instantiate
each of these components, providing them with a simple interface to
issue commands and subscribe to changes:

    {
      playMidiNote,
      stopMidiNote,
      onMidiNotePlayed,
      onMidiNoteStopped,
    }

which can be extended to other commands/events e.g. drawbar changes.

At the moment all the top-level controller gubbins is still happening in
main.js, but I can extract that later.