Extract ManualKeys const for Synth
I'm trying to decouple the various components of this system, and one
coupling we have going on is that building the HTML UI requires knowing
what the range of keys is for the manual.

Rather than accessing this via a method on the Synth object, this
extracts a const that can be imported directly. For the time being I'm
leaving this in the Synth module, but probably this (along with the tone
wheels) will end up in a component-agnostic definitions module at some
point.