Extract definitions and stop mutating ToneWheels
I'd like the properties of the hammond (e.g. tonewheels, manuals,
drawbars etc.) to be statically defined and importable by any component
that needs them.

This extracts the ManualKeys definition to a shared definitions module,
and also stops the Synth module from mutating the ToneWheels array; it
doesn't need to do this, once it's built the oscillators and gain nodes
for each wheel it's done with the definition info.