Sum tonewheel gain, don't overwrite it
Now that I can play lots of notes at once, it's obvious that for each
note being played, I'm overwriting the gain for each tonewheel that note
voices (depending on the drawbar settings).

Instead, I should be summing the gain for each tonewheel across all
played notes that voice it.

This has introduced some distortion when all the stops are out, so in
the next commit I'll normalise the main gain each time the drawbar
settings are changed.