Normalise synth main gain
I'd like approximately equivalent volume after changing the drawbar
settings, so this applies a normalisation every time they change.

I may wire in a master volume control but for now 0.5 / (sum of
drawbars) seems to result in an okay level that doesn't distort if you
press all the keys on my midi controller.