Separate drawbar definitions from current values
This is part of the ongoing component decoupling; in this commit I'm
trying to make the drawbar UI construction less dependent on the synth
module by extracting the static Drawbars definition and using this to
build the UI instead of synth.eachDrawbar.

There's a little wonk here in that the UI is now built without knowledge
of the initial drawbar values, and it's really just happenstance that
the UI matches the synth settings (all stops pulled out). I'll worry
about that later.