Use controller events to play notes via Synth
This is the first introduction of event handlers for a top-level
controller object, which I'll be using to update all components whenever
a change takes place elsewhere in the system. Currently only Synth uses
it, but soon it'll be used for a yet-to-be-introduced UI module (whose
"output" in response to events is e.g. showing a given key as being
pressed).