Sort out drawbar input vs output
This approximately fixes it so that on page load the UI actually gets
set to what the drawbar values are, instead of accidentally matching
them.

I'm still not super happy with this but I can't put my finger on why
right now. I think it's because it's not clear where the responsibility
for the initial config of the synthesiser should live, and I guess
implicitly now it's living at the main module level. Which is fine I
suppose. I don't love the way it's injected via mutations though. Maybe
there should be an initial state that's passed to the different
components, and only after initialisation do they start communicating
via controller events?

Well anyway. Problems for future Simon, I want to build more features
now so I'm calling time on the refactoring for now.