Working percussion effect with hard-coded note and delay
The hammond percussion effect, when enabled, applies a single-triggered
attack envelope to a note on the second or third harmonics of the
fundamental / 4' or 2 2/3' drawbars.

Single-triggered means that only the first note played from a
fully-released keyboard will sound the percussion effect.
Thereafter, as long as any key is being held, no new notes will have the
percussion effect applied.

In practice, there's some slop around this to enable percussion on
chords - as long as you get a group of keys down at approximately the
same time from a fully-released state, you'll get the percussion effect
on all of them.

To implement this I've added a latch mechanism with a short amount of
slop to it; any key pressed within 50ms of the first key will also get
percussion, after that the whole latch must be released again for any
further percussion effects to sound.

In terms of sounding the actual effect, I've hard-coded both the
amplitude and note for the time being; controls to follow shortly.

I'm using linear ramps for both the attack and decay of the percussion
rather than the exponential setTargetAtTime, more via trial and error
than anything else.

Doing the attack fairly rapidly does lead to that sort of clicking
effect, but having watched a few youtube videos of slightly odd fellas
demoing the percussion effect, I think that's not unrealistic, so I'm
leaving it there for now. Once I've got basic versions of all the
effects in I can do a second pass for realism. Maybe.

https://www.youtube.com/watch?v=TCEWw7StW1o