Don't repeatedly play held keys
With the computer keyboard input, key repeat means you get multiple
keydown events without a keyup one; this isn't a problem when notes are
just on or off, but now we have percussion there's a specific keydown
effect, so we only want to support playing notes that aren't already
being played.

I could've implemented this in the keyboard input module but it doesn't
seem unreasonable to put it in the synth, which should support only
operations it's actually possible to do with a hammond manual. Also the
state about what's currently playing is maintained here, and I'm not
desperate to make the keyboard input stateful as well.