Stop normalizing main gain based on drawbars
This seemed like a good idea at the time, but for testing percussion
effects it's useful to be able to shut down all the drawbars so only the
percussion is voiced, and that produces a divide by zero.

I should really just plug in a master volume control into the UI, but
for the time being this reverts to a fixed main gain.