Basic vibrato with hard-coded depth and frequency
The hammond vibrato is implemented as a segmented delay line[1], which
is tapped at various points to expose contacts with different degrees of
delay. A rotating scanner arm then passes over these contacts to vary
the amount of delay introduced over time, and this in turn alters the
frequency of the signal on the line.

I've implemented this here in what seemed like the obvious way, with a
DelayNode providing the, y'know, delay, and a low-frequency oscillator
(LFO) and a gain node driving the amount of delay. The frequency of the
LFO determines the pace of the vibrato, and a combination of this and
the gain determine the depth the resulting delta in signal frequency,
a.k.a. depth.

For the moment I've hard-coded these values and am using a pure sine
wave for the LFO. Based on [2] the Hammond's own delay scanner
introduces a max 500µs of delay, so I've set it (I think!) to a midpoint
of 250µs, a max of 500µs, and set the gain to .00025 in the hope that
this results in a sweep from 0 to 500µs. I'm not certain this 500µs is
correct, as [3] states the max delay is 1ms, but I'll start here and can
adjust later.

According to [3, 4] the B-3 vibrato has no pace control, and is driven
at a constant pace of "around 7Hz"[3] or 412 RPM[4], which is 6.87Hz, so
I'm happy those two corroborate each other.

For the time being I'm setting the LFO as a sine wave, but in practice
the waveform of the delay line modulation is more complex. [3] says the
delay line is 9-stage and that the scanner sweeps back and forth across
these linearly, resulting in a triangular wave of delay modulation.

However [4, 5] refer to an 18-stage delay line, and a much more complex
system of selecting how the different settings tap the line at different
points in the sweep, to produce something closer to a sine wave of delay
modulation. I think these sources seem more authoritative so I'm going
to follow them in later improvements, and implement three custom
waveforms for the LFO that try to mimic the delay properties of the real
scanner.

[1] https://bentonelectronics.com/service-manual-the-hammond-vibrato/
[2] https://modularsynthesis.com/modules/DJB-scanner/scanner.htm
[3] https://electricdruid.net/index.php?page=info.hammond#scanner
[4] https://b3world.com/hammond-technical-information-02.html
[5] http://www.stefanv.com/electronics/hammond_vibrato_mod.html