Rejig VibratoNode to support multiple internal paths
The vibrato / chorus module's chorus effect is (to my understanding)
achieved by mixing the output of the delay line with the unmodified
source.

To do this within the VibratoNode (i.e. transparently to the consumer) I
need to support multiple paths from its input to its output. However as
originally implemented, the DelayNode is effectively the input node, so
we only get access to the delayed source.

This switches us to use a GainNode with unity gain as the base node, and
we can then connect this on construction to as many internal paths as we
like.

Note that we can't just override the connect() method in this subclass
now, because we need to use the GainNode's original connect()
implementation in the constructor, and only then switch out the public
connect() interface so that it wires up the VibratoNode's internals to
the output.

In the next commit I'll add the chorus line. Then finally I guess I'll
wire up some config for the different settings of the vibrato module.