WIP scrolling drawbars seems hard
Picking up the scroll event doesn't work because that keys off the
actual movement of the element, which we want to control (and thus lock
down with `scroll: hidden`).

The wheel event seems super inconsistent though (at least on the macbook
trackpad) and when gently used just fires a lot of zero-delta events.
I'm leaving this here for the moment but it needs a lot of work to be
even nearly pleasing.

An outright bug in my implementation is that continuing to scroll "past"
the stops at each end still adds to the `this.wheeled` total, so you
need to then scroll back to where you were for the thing to move again.
Should be an easy enough fix but the behaviour is just so wonky I can't
be bothered and am going to do something else more fun instead.