Implement all vibrato modes and add form
This adds settings and UI controls for the hammond vibrato modes. The V-
modes are pure vibrato, while the C- modes include a mixture of the
vibrato line and the unmodified input for a chorus effect.

The numbers indicate (I believe!) the vibrato depth; I'm still
researching what actual depth and waveform these produce, so for the
time being I've implemented it in even one-third steps of the amplitude
applied to the vibrato LFO.

I think the different event handlers and synth control setters are
getting out of hand here, and I'm a bit worried about the data flow
directions on these settings, so in subsequent commits I think I'm going
to turn the UI into one big form whose config can be sent en masse to
the synth any time a setting changes. This would also mean I'd more
easily be able to grab the form's initial state and use that as the
canonical reference for the initial synth settings.

I persist with the JS templating for reasons best known to myself. Maybe
it'd just be easier all as static HTML, or driven some other way. We'll
see I guess.