Try to sort out safe audioContext initialization
The page is currently broken if you haven't blanket-allowed audio for
it, because now that we have settings that run on initialization (e.g.
the vibrato setting) that modify audio nodes, if the audio context isn't
initialized at that point, the operation will fail. And since this
happens before the user's had a chance to make any page input, it almost
certainly will.

I'm not really happy with this solution but it means we don't try and
boot up any of the synth at all until we know we've got a running audio
context.

Maybe later I'll make the synth module better able to recover from an
initially-suspended context, but since midi keypresses don't seem to
count as user input for the purpose of resuming it, this hacky "just
have a big button" approach seems fine for now. Maybe I'll gussy it up
by styling it like a power switch, idk.