Set up QUnit and add an example unit test
I decided to let myself off the "no dependencies" rule for testing
only, because I don't want to write (even a tiny) testing framework.

BUT I'm still not using any node modules or build step, it's purely
vendored and in-browser. So I think this is still in the spirit of
things.