Rework page layout a bit
I want to think about making this responsive (it's currently not at
all), which this commit doesn't do at all but does switch to an overall
flex layout I think it'll be easier to work within, and centres
everything. I'm also making the whole page brown cos the white-on-brown
rectangle didn't look great.