Don't make synth depend on controller
Coming back to this after ~2 years it's barking to have an ersatz event
emitter sitting in the middle of everything, ultimately just doing
spooky action-at-a-distance dispatch of function calls to the synth.

Yes, it'll be useful to have events emitted so the UI can update itself
but I can just have the synth be an EventEmitter and dispense with this
whole controller object. Then the input classes can call the synth
directly and we're altogether much simpler.