Eliminate Controller by making Synth an EventTarget
This gets rid of the needless indirection and makes me feel better about
things in general. Synth control actions feel like they should be
synchronous, while it still (to me) makes sense to have the visual
keyboard display respond to changes in synth state subscribed to via
event.

I think I'd prefer to have a single synth state change event to
subscribe to, which would rely on a firmer representation of the current
synth state, which would be good to have anyway.